(init) module organization
