+++ src/init.lua
@@ -22,16 +22,32 @@ Logo: figlet -f soft Theovim
 "                                                                    "
 --]]
 --
+
 -- Lua configs in ~/.config/nvim/lua
-require("theo_init")
-require("plugins")
-require("look")
+-- Remove a hyphen before brackets to disable a module
 
-require("file_et_search")
-require("lsp")
+---[[ Core
+require("core/theo_init")
+require("core/keybindings")
+require("core/plugins")
+--]]
 
-require("theovim")
-require("custom_menus")
+---[[ Look
+require("look/look")
+require("look/statusline")
+--]]
 
-local status, user_config_call = pcall(require, "user_config")
+---[[ Editor
+require("editor/lsp")
+require("editor/file_et_search")
+--]]
+
+---[[ Theovim
+require("theovim/theovim")
+require("theovim/custom_menus")
+--]]
+
+---[[ Safeguards around including user configuration file
+local status, _ = pcall(require, "user_config")
 if (not status) then return end
+--]]
