+++ src/lua/core/plugins.lua
